requirements.txt
albumentations==1.3.1
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.41.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.23.4
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
sentry-sdk==1.29.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.7.18
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.7
